Chore(ostor): refactor (split key methods)
