+++ pkg/ostor/user.go
@@ -1,9 +1,5 @@
 package ostor
 
-import (
-	"github.com/go-resty/resty/v2"
-)
-
 // query parameter for user management
 const qUsers string = "ostor-users"
 
@@ -23,14 +19,4 @@ func (o *Ostor) GetUser(email string) (*OstorUser, error) {
 	var user *OstorUser
 	_, err := o.get(qUsers, map[string]string{"emailAddress": email}, &user)
 	return user, err
-}
-
-func (o *Ostor) GenerateCredentials(email string) (*OstorCreateUserResponse, error) {
-	var user *OstorCreateUserResponse
-	_, err := o.post(qUsers, qUsers+"&emailAddress="+email+"&genKey", user)
-	return user, err
-}
-
-func (o *Ostor) RevokeKey(email, accessKeyID string) (*resty.Response, error) {
-	return o.post(qUsers, qUsers+"&emailAddress="+email+"&revokeKey="+accessKeyID, nil)
 }
